backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlmodel
asyncpg
python-dotenv
alembic
pytest
pytest-asyncio
httpx
passlib[argon2]
aiosqlite
python-jose[cryptography]
python-multipart
email-validator
